+++ lithops/scripts/cli.py
@@ -65,7 +65,6 @@ from lithops.storage import InternalStorage
 from lithops.serverless import ServerlessHandler
 from lithops.storage.utils import clean_bucket
 from lithops.standalone import StandaloneHandler
-from lithops.standalone.utils import StandaloneMode
 from lithops.localhost import LocalhostHandler
 
 
@@ -780,7 +779,6 @@ def build_image(ctx, name, file, config, backend, region, debug, overwrite):
 
     config = load_yaml_config(config) if config else None
     config_ow = set_config_ow(backend=backend, region=region)
-    config_ow['backend']['exec_mode'] = StandaloneMode.CREATE.value
     config = default_config(config_data=config, config_overwrite=config_ow, load_storage_config=False)
 
     if config['lithops']['mode'] != STANDALONE:
@@ -810,7 +808,6 @@ def delete_image(ctx, name, config, backend, region, debug):
 
     config = load_yaml_config(config) if config else None
     config_ow = set_config_ow(backend=backend, region=region)
-    config_ow['backend']['exec_mode'] = StandaloneMode.CREATE.value
     config = default_config(config_data=config, config_overwrite=config_ow, load_storage_config=False)
 
     if config['lithops']['mode'] != STANDALONE:
@@ -836,7 +833,6 @@ def list_images(config, backend, region, debug):
 
     config = load_yaml_config(config) if config else None
     config_ow = set_config_ow(backend=backend, region=region)
-    config_ow['backend']['exec_mode'] = StandaloneMode.CREATE.value
     config = default_config(config_data=config, config_overwrite=config_ow, load_storage_config=False)
 
     if config['lithops']['mode'] != STANDALONE: